representation/polars_to_rdf: Move language column extraction to function
