+++ lib/representation/src/polars_to_rdf.rs
@@ -1,5 +1,5 @@
 use crate::errors::RepresentationError;
-use crate::multitype::{all_multi_main_cols, MULTI_BLANK_DT, MULTI_IRI_DT, MULTI_NONE_DT};
+use crate::multitype::{all_multi_main_cols, extract_lang_literal_from_multitype, MULTI_BLANK_DT, MULTI_IRI_DT, MULTI_NONE_DT};
 use crate::rdf_to_polars::{
     polars_literal_values_to_series, rdf_literal_to_polars_literal_value,
     rdf_owned_blank_node_to_polars_literal_value, rdf_owned_named_node_to_polars_literal_value,
@@ -49,37 +49,7 @@ pub fn column_as_terms(column: &Column, t: &RDFNodeType) -> Vec<Option<Term>> {
             let mut iters: Vec<IntoIter<Option<Term>>> = vec![];
             for (t, colname) in types.iter().zip(all_multi_main_cols(types)) {
                 let v = if t.is_lang_string() {
-                    let mut lf = DataFrame::new(vec![
-                        column
-                            .struct_()
-                            .unwrap()
-                            .field_by_name(LANG_STRING_VALUE_FIELD)
-                            .unwrap()
-                            .cast(&DataType::String)
-                            .unwrap()
-                            .clone()
-                            .into_column(),
-                        column
-                            .struct_()
-                            .unwrap()
-                            .field_by_name(LANG_STRING_LANG_FIELD)
-                            .unwrap()
-                            .cast(&DataType::String)
-                            .unwrap()
-                            .clone()
-                            .into_column(),
-                    ])
-                    .unwrap()
-                    .lazy();
-                    lf = lf.with_column(
-                        as_struct(vec![
-                            col(LANG_STRING_LANG_FIELD),
-                            col(LANG_STRING_VALUE_FIELD),
-                        ])
-                        .alias(&colname),
-                    );
-                    let df = lf.collect();
-                    let ser = df.unwrap().drop_in_place(&colname).unwrap();
+                    let ser = extract_lang_literal_from_multitype(column, colname);
                     basic_rdf_node_type_column_to_term_vec(&ser, t)
                 } else {
                     basic_rdf_node_type_column_to_term_vec(
